Enhance feature flag processing and clarify provider test instructions
